create form.js
